+++ src/main/kotlin/io/openfuture/chain/core/model/entity/transaction/unconfirmed/UnconfirmedDelegateTransaction.kt
@@ -1,8 +1,8 @@
 package io.openfuture.chain.core.model.entity.transaction.unconfirmed
 
+import io.openfuture.chain.core.model.entity.transaction.TransactionFooter
 import io.openfuture.chain.core.model.entity.transaction.TransactionHeader
 import io.openfuture.chain.core.model.entity.transaction.payload.DelegateTransactionPayload
-import io.openfuture.chain.core.model.entity.transaction.payload.TransactionPayload
 import io.openfuture.chain.network.message.core.DelegateTransactionMessage
 import io.openfuture.chain.rpc.domain.transaction.request.DelegateTransactionRequest
 import javax.persistence.Embedded
@@ -13,42 +13,34 @@ import javax.persistence.Table
 @Table(name = "u_delegate_transactions")
 class UnconfirmedDelegateTransaction(
     header: TransactionHeader,
-    hash: String,
-    senderSignature: String,
-    senderPublicKey: String,
+    footer: TransactionFooter,
 
     @Embedded
     var payload: DelegateTransactionPayload
 
-) : UnconfirmedTransaction(header, hash, senderSignature, senderPublicKey) {
+) : UnconfirmedTransaction(header, footer) {
 
     companion object {
         fun of(message: DelegateTransactionMessage): UnconfirmedDelegateTransaction = UnconfirmedDelegateTransaction(
             TransactionHeader(message.timestamp, message.fee, message.senderAddress),
-            message.hash,
-            message.senderSignature,
-            message.senderPublicKey,
+            TransactionFooter(message.hash, message.senderSignature, message.senderPublicKey),
             DelegateTransactionPayload(message.delegateKey, message.delegateHost, message.delegatePort)
         )
 
         fun of(request: DelegateTransactionRequest): UnconfirmedDelegateTransaction = UnconfirmedDelegateTransaction(
             TransactionHeader(request.timestamp!!, request.fee!!, request.senderAddress!!),
-            request.hash!!,
-            request.senderSignature!!,
-            request.senderPublicKey!!,
+            TransactionFooter(request.hash!!, request.senderSignature!!, request.senderPublicKey!!),
             DelegateTransactionPayload(request.delegateKey!!, request.senderHost!!, request. senderPort!!)
         )
     }
-
-    override fun getPayload(): TransactionPayload = payload
 
     override fun toMessage(): DelegateTransactionMessage = DelegateTransactionMessage (
         header.timestamp,
         header.fee,
         header.senderAddress,
-        hash,
-        senderSignature,
-        senderPublicKey,
+        footer.hash,
+        footer.senderSignature,
+        footer.senderPublicKey,
         payload.delegateKey,
         payload.delegateHost,
         payload.delegatePort