+++ src/main/kotlin/io/openfuture/chain/core/model/entity/transaction/unconfirmed/UnconfirmedVoteTransaction.kt
@@ -1,7 +1,7 @@
 package io.openfuture.chain.core.model.entity.transaction.unconfirmed
 
+import io.openfuture.chain.core.model.entity.transaction.TransactionFooter
 import io.openfuture.chain.core.model.entity.transaction.TransactionHeader
-import io.openfuture.chain.core.model.entity.transaction.payload.TransactionPayload
 import io.openfuture.chain.core.model.entity.transaction.payload.VoteTransactionPayload
 import io.openfuture.chain.network.message.core.VoteTransactionMessage
 import io.openfuture.chain.rpc.domain.transaction.request.VoteTransactionRequest
@@ -13,42 +13,34 @@ import javax.persistence.Table
 @Table(name = "u_vote_transactions")
 class UnconfirmedVoteTransaction(
     header: TransactionHeader,
-    hash: String,
-    senderSignature: String,
-    senderPublicKey: String,
+    footer: TransactionFooter,
 
     @Embedded
     val payload: VoteTransactionPayload
 
-) : UnconfirmedTransaction(header, hash, senderSignature, senderPublicKey) {
+) : UnconfirmedTransaction(header, footer) {
 
     companion object {
         fun of(message: VoteTransactionMessage): UnconfirmedVoteTransaction = UnconfirmedVoteTransaction(
             TransactionHeader(message.timestamp, message.fee, message.senderAddress),
-            message.hash,
-            message.senderSignature,
-            message.senderPublicKey,
+            TransactionFooter(message.hash, message.senderSignature, message.senderPublicKey),
             VoteTransactionPayload(message.voteTypeId, message.delegateKey)
         )
 
         fun of(request: VoteTransactionRequest): UnconfirmedVoteTransaction = UnconfirmedVoteTransaction(
             TransactionHeader(request.timestamp!!, request.fee!!, request.senderAddress!!),
-            request.hash!!,
-            request.senderSignature!!,
-            request.senderPublicKey!!,
+            TransactionFooter(request.hash!!, request.senderSignature!!, request.senderPublicKey!!),
             VoteTransactionPayload(request.voteType!!.getId(), request.delegateKey!!)
         )
     }
-
-    override fun getPayload(): TransactionPayload = payload
 
     override fun toMessage(): VoteTransactionMessage = VoteTransactionMessage(
         header.timestamp,
         header.fee,
         header.senderAddress,
-        hash,
-        senderSignature,
-        senderPublicKey,
+        footer.hash,
+        footer.senderSignature,
+        footer.senderPublicKey,
         payload.voteTypeId,
         payload.delegateKey
     )