+++ src/main/kotlin/io/openfuture/chain/core/service/transaction/DefaultTransactionService.kt
@@ -1,6 +1,7 @@
 package io.openfuture.chain.core.service.transaction
 
 import io.openfuture.chain.core.component.TransactionCapacityChecker
+import io.openfuture.chain.core.exception.NotFoundException
 import io.openfuture.chain.core.model.entity.transaction.confirmed.Transaction
 import io.openfuture.chain.core.model.entity.transaction.unconfirmed.UnconfirmedTransaction
 import io.openfuture.chain.core.repository.TransactionRepository
@@ -22,6 +23,10 @@ class DefaultTransactionService(
 
     @Transactional(readOnly = true)
     override fun getCount(): Long = repository.count()
+
+    @Transactional(readOnly = true)
+    override fun getUnconfirmedTransactionByHash(hash: String): UnconfirmedTransaction = uRepository.findOneByFooterHash(hash)
+        ?: throw NotFoundException("Unconfirmed transaction with hash $hash not found")
 
     override fun getProducingPerSecond(): Long {
         return capacityChecker.getCountPerSecond()