+++ src/main/kotlin/io/openfuture/chain/rpc/domain/block/BaseBlockResponse.kt
@@ -4,7 +4,6 @@ abstract class BaseBlockResponse(
     val timestamp: Long,
     val height: Long,
     val previousHash: String,
-    val reward: Long,
     val hash: String,
     val signature: String,
     val publicKey: String