+++ src/main/kotlin/io/openfuture/chain/rpc/domain/block/MainBlockResponse.kt
@@ -6,25 +6,23 @@ class MainBlockResponse(
     timestamp: Long,
     height: Long,
     previousHash: String,
-    reward: Long,
     hash: String,
     signature: String,
     publicKey: String,
     val merkleHash: String,
-    val transactionsCount: Long,
+    val transactionsCount: Int,
     val epochIndex: Long
-) : BaseBlockResponse(timestamp, height, previousHash, reward, hash, signature, publicKey) {
+) : BaseBlockResponse(timestamp, height, previousHash, hash, signature, publicKey) {
 
     constructor(block: MainBlock, epochIndex: Long) : this(
         block.timestamp,
         block.height,
         block.previousHash,
-        block.reward,
         block.hash,
         block.signature,
         block.publicKey,
         block.payload.merkleHash,
-        (block.payload.transferTransactions.size + block.payload.voteTransactions.size + block.payload.delegateTransactions.size).toLong(),
+        block.getTransactionsCount(),
         epochIndex
     )