+++ src/test/kotlin/io/openfuture/chain/component/block/DefaultPendingBlockHandlerTests.kt
@@ -12,6 +12,7 @@ import io.openfuture.chain.core.service.MainBlockService
 import io.openfuture.chain.crypto.util.SignatureUtils
 import io.openfuture.chain.network.message.consensus.BlockApprovalMessage
 import io.openfuture.chain.network.message.consensus.PendingBlockMessage
+import io.openfuture.chain.network.message.core.RewardTransactionMessage
 import io.openfuture.chain.network.service.NetworkApiService
 import org.bouncycastle.pqc.math.linearalgebra.ByteUtils
 import org.junit.Before
@@ -50,9 +51,9 @@ class DefaultPendingBlockHandlerTests : ServiceTests() {
         )
 
         val privateKey = "529719453390370201f3f0efeeffe4c3a288f39b2e140a3f6074c8d3fc0021e6"
-
-        val pendingBlock = PendingBlockMessage(2L, "previousHash", 1L, 2L,
-            "hash", "signature", "publicKey", payload.merkleHash, listOf(), listOf(), listOf())
+        val rewardTransactionMessage = createRewardTransactionMessage(1L)
+        val pendingBlock = PendingBlockMessage(2L, "previousHash", 1L,
+            "hash", "signature", "publicKey", payload.merkleHash, rewardTransactionMessage, listOf(), listOf(), listOf())
 
         given(keyHolder.getPrivateKey()).willReturn(
             ByteUtils.fromHexString(privateKey))
@@ -78,9 +79,9 @@ class DefaultPendingBlockHandlerTests : ServiceTests() {
         val payload = MainBlockPayload(
             "merkleHash"
         )
-
-        val pendingBlock = PendingBlockMessage(2L, "previousHash", 1L, 2L,
-            hash, "signature", publicKey, payload.merkleHash, listOf(), listOf(), listOf())
+        val rewardTransactionMessage = createRewardTransactionMessage(1L)
+        val pendingBlock = PendingBlockMessage(2L, "previousHash", 1L,
+            hash, "signature", publicKey, payload.merkleHash, rewardTransactionMessage, listOf(), listOf(), listOf())
 
         val message = BlockApprovalMessage(
             BlockApprovalStage.PREPARE.value,
@@ -156,5 +157,10 @@ class DefaultPendingBlockHandlerTests : ServiceTests() {
 
         verify(networkService, times(2)).broadcast(any(BlockApprovalMessage::class.java))
     }
+
+    private fun createRewardTransactionMessage(timestamp: Long): RewardTransactionMessage =
+        RewardTransactionMessage(timestamp, 0, "senderAddress", "hash",
+            "senderSignature", "senderPublicKey", 10,
+            "recipientAddress")
 
 }