+++ src/test/kotlin/io/openfuture/chain/rpc/controller/DelegateControllerTests.kt
@@ -48,7 +48,7 @@ class DelegateControllerTests : ControllerTests() {
     @Test
     fun getAllActiveShouldReturnActiveDelegatesListTest() {
         val delegate = Delegate("publicKey", "address", "host", 1)
-        val genesisBlock = GenesisBlock(1, 1, "previousHash", 1, "hash", "signature", "publicKey",
+        val genesisBlock = GenesisBlock(1, 1,   "previousHash", "hash", "signature", "publicKey",
             GenesisBlockPayload(1, listOf(delegate)))
         val expectedPageResponse = PageResponse(PageImpl(listOf(DelegateResponse(delegate))))