+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - RPC: Main and genesis block endpoints unit tests
 - Wallet UI: Create a transaction
 - Wallet UI: Past transaction list
+- Core: Reward transaction
 
 ### Changed
 - Core: Genesis and Main blocks are not broadcasted after creation, pending blocks are broadcasted instead
@@ -31,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Core validation: Improved fee and amount validation for the transfer transaction
 - Core: Removed validation from main block service on adding a block
 - Core: Removed validation from genesis service on adding a block
+- Core: Removed field 'reward' from blocks
 
 ## [0.5.0] - 2018-08-10
 ### Added